# docs/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080/walkbg
  description: Generated server url
...
paths:
...
  /api/user:
    post:
      tags:
      - 用户管理
      summary: 创建用户
      description: 创建新的用户
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseUserBasicResponse'
...
  /api/user/{id}:
    get:
      tags:
      - 用户管理
      summary: 根据ID获取用户
      description: 根据用户ID获取用户信息
      operationId: getUserById
      parameters:
      - name: id
        in: path
        description: 用户ID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseUserBasicResponse'
  /api/user/username/{username}:
    get:
      tags:
      - 用户管理
      summary: 根据用户名获取用户
      description: 根据用户名获取用户信息
      operationId: getUserByUsername
      parameters:
      - name: username
        in: path
        description: 用户名
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponseUserBasicResponse'
  /api/trips/user/{userId}:
    get:
      tags:
      - trip-controller
      operationId: getUserTrips
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageTrip'
...
  /api/trip-participants/user/{userId}:
    get:
      tags:
      - trip-participant-controller
      operationId: getUserTripParticipations
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageTripParticipant'
...
  /api/trip-participants/recent/user/{userId}:
    get:
      tags:
      - trip-participant-controller
      operationId: getRecentUserParticipations
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
      - name: pageable
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageTripParticipant'
  /api/trip-participants/count/user/{userId}:
    get:
      tags:
      - trip-participant-controller
      operationId: countUserTripParticipations
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
                additionalProperties:
                  type: integer
                  format: int64
...
components:
  schemas:
...
    Trip:
      required:
      - createdAt
      - id
      - imageUrls
      - name
      - organizerId
      - privacySetting
      - routeIds
      - status
      - updatedAt
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        organizerId:
          type: string
        primaryRouteId:
          type: string
        budget:
          type: number
        actualCost:
          type: number
        notes:
          type: string
        privacySetting:
          type: integer
          format: int32
        coverUrl:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        imageUrls:
          type: array
          items:
            type: string
        routeIds:
          type: array
          items:
            type: string
...
    UserCreateRequest:
      required:
      - email
      - username
      type: object
      properties:
        username:
          maxLength: 50
          minLength: 3
          pattern: "^[a-zA-Z0-9_]+$"
          type: string
        email:
          maxLength: 100
          minLength: 0
          type: string
        nickname:
          maxLength: 50
          minLength: 0
          type: string
        phone:
          pattern: "^[1-9]\\d{10}$"
          type: string
        avatar_url:
          maxLength: 500
          minLength: 0
          pattern: ^(https?://.*|)$
          type: string
    ApiResponseUserBasicResponse:
      required:
      - code
      - message
      - success
      - timestamp
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          $ref: '#/components/schemas/UserBasicResponse'
        timestamp:
          type: string
          format: date-time
        code:
          type: integer
          format: int32
    UserBasicResponse:
      required:
      - created_at
      - email
      - id
      - username
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        nickname:
          type: string
        phone:
          type: string
        avatar_url:
          type: string
        created_at:
          type: integer
          format: int64
    TripParticipant:
      required:
      - joinedAt
      - role
      - status
      - tripId
      - userId
      type: object
      properties:
        tripId:
          type: string
        userId:
          type: string
        role:
          type: integer
          format: int32
        status:
          type: integer
          format: int32
        joinedAt:
          type: string
          format: date-time
...
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/SortObject'
        paged:
          type: boolean
        unpaged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
...
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
...
    PageTrip:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/Trip'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        pageable:
          $ref: '#/components/schemas/PageableObject'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        empty:
          type: boolean
    PageTripParticipant:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/TripParticipant'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        pageable:
          $ref: '#/components/schemas/PageableObject'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        empty:
          type: boolean
...